# src/app/api/runs/[id]/route.ts
import { NextResponse, NextRequest } from 'next/server';
import { getRunById } from '@/lib/runService';
import { z } from 'zod';
import { getServerSession } from 'next-auth/next';
import { authOptions } from '@/app/api/auth/[...nextauth]/route';
import { updateRun, UpdateRunData } from '@/lib/runService';
import { PrismaClientKnownRequestError } from '@/generated/prisma/runtime/library';

// Schema to validate the ID from the path
const paramsSchema = z.object({
  id: z.string().cuid({ message: 'Invalid run ID format' }),
});

interface RouteContext {
  params: {
    id: string;
  };
}

export async function GET(request: NextRequest, context: RouteContext) {
  // No session check needed for viewing a single run as per current requirements (publicly viewable)
  // If authentication is required later, add session check here.

  // console.log('Type of context.params:', typeof context.params); // Keep for now if needed
  // console.log('Value of context.params:', context.params); // Keep for now if needed

  // Await context.params itself if it's promise-like, then access id
  const resolvedParams = await context.params;
  const runIdFromContext = resolvedParams.id;
  const paramsToValidate = { id: runIdFromContext };

  const validationResult = paramsSchema.safeParse(paramsToValidate);

  if (!validationResult.success) {
    console.error(
      'Invalid ID Validation Error (GET):',
      validationResult.error.flatten()
    );
    return NextResponse.json(
      {
        message: 'Invalid run ID format',
        errors: validationResult.error.flatten().fieldErrors,
      },
      { status: 400 }
    );
  }

  const { id } = validationResult.data;

  try {
    const run = await getRunById(id);

    if (!run) {
      return NextResponse.json({ message: 'Run not found' }, { status: 404 });
    }

    return NextResponse.json(run, { status: 200 });
  } catch (error) {
    console.error(`Error fetching run with ID ${id}:`, error);
    return NextResponse.json(
      { message: 'Internal Server Error' },
      { status: 500 }
    );
  }
}

// Zod schema for PUT request body (all fields optional for partial updates)
const updateRunSchema = z.object({
  number: z.number().int().positive().optional(),
  descriptor: z
    .string()
    .min(3, { message: 'Descriptor must be at least 3 characters long' })
    .optional(),
  dateTime: z
    .string()
    .datetime({ message: 'Invalid datetime string. Must be ISO8601' })
    .optional(),
  address: z
    .string()
    .min(5, { message: 'Address must be at least 5 characters long' })
    .optional(),
  lat: z.number().nullable().optional(), // Allow null to unset
  lng: z.number().nullable().optional(), // Allow null to unset
  introLink: z
    .string()
    .url({ message: 'Invalid URL for intro link' })
    .nullable()
    .optional()
    .or(z.literal('')), // Allow null or empty string to unset
});

export async function PUT(request: NextRequest, context: RouteContext) {
  const session = await getServerSession(authOptions);

  if (!session || !session.user || !session.user.id) {
    return NextResponse.json({ message: 'Unauthorized' }, { status: 401 });
  }

  // Authorization: Check if user has 'organizer' or 'admin' role
  // @ts-ignore // session.user.role might not be typed by default in NextAuth
  if (session.user.role !== 'ORGANIZER' && session.user.role !== 'ADMIN') {
    return NextResponse.json(
      { message: 'Forbidden: Insufficient permissions' },
      { status: 403 }
    );
  }

  // Await context.params itself if it's promise-like, then access id for PUT
  const resolvedParamsPut = await context.params;
  const runIdFromContextPut = resolvedParamsPut.id;
  const paramsToValidatePut = { id: runIdFromContextPut };

  const paramsValidationResult = paramsSchema.safeParse(paramsToValidatePut);

  if (!paramsValidationResult.success) {
    console.error(
      'Invalid ID Validation Error (PUT):',
      paramsValidationResult.error.flatten()
    );
    return NextResponse.json(
      {
        message: 'Invalid run ID format',
        errors: paramsValidationResult.error.flatten().fieldErrors,
      },
      { status: 400 }
    );
  }
  const { id } = paramsValidationResult.data; // Use the validated id from paramsToValidatePut

  try {
    const body = await request.json();
    const validationResult = updateRunSchema.safeParse(body);

    if (!validationResult.success) {
      return NextResponse.json(
        {
          message: 'Invalid input',
          errors: validationResult.error.flatten().fieldErrors,
        },
        { status: 400 }
      );
    }

    const updateData: UpdateRunData = {};
    const validatedBody = validationResult.data;

    // Construct the updateData object, only including fields that were actually provided
    if (validatedBody.number !== undefined)
      updateData.number = validatedBody.number;
    if (validatedBody.descriptor !== undefined)
      updateData.descriptor = validatedBody.descriptor;
    if (validatedBody.dateTime !== undefined)
      updateData.dateTime = new Date(validatedBody.dateTime);
    if (validatedBody.address !== undefined)
      updateData.address = validatedBody.address;
    if (validatedBody.lat !== undefined) updateData.lat = validatedBody.lat; // Can be null
    if (validatedBody.lng !== undefined) updateData.lng = validatedBody.lng; // Can be null
    if (validatedBody.introLink !== undefined) {
      updateData.introLink =
        validatedBody.introLink === '' ? null : validatedBody.introLink; // Convert empty string to null
    }

    // Prevent updating with an empty object if no valid fields were passed
    if (Object.keys(updateData).length === 0) {
      return NextResponse.json(
        { message: 'No valid fields provided for update' },
        { status: 400 }
      );
    }

    const updatedRun = await updateRun(id, updateData);

    if (!updatedRun) {
      return NextResponse.json(
        { message: 'Run not found or update failed' },
        { status: 404 }
      );
    }

    return NextResponse.json(updatedRun, { status: 200 });
  } catch (error) {
    console.error(`Error updating run with ID ${id}:`, error);
    if (error instanceof PrismaClientKnownRequestError) {
      if (error.code === 'P2002' && error.meta?.target === 'Run_number_key') {
        return NextResponse.json(
          { message: 'A run with this number already exists.' },
          { status: 409 }
        );
      }
    }
    return NextResponse.json(
      { message: 'Internal Server Error' },
      { status: 500 }
    );
  }
}